requirements.txt
nltk~=3.6.2
natasha~=1.4.0
pyenchant~=3.2.0
pymorphy2~=0.9.1
SQLAlchemy~=1.4.15
Werkzeug~=1.0.1
pytesseract~=0.3.7
WTForms~=2.3.3
Flask~=1.1.2
requests~=2.25.1
dominate~=2.6.0
flask_httpauth